+++ t/cases/794_mb611_achievement_thresholds.t
@@ -80,6 +80,8 @@ return sub {
         [ 'nuit',      qw(night_owl midnight_regular creature_night witching_hour) ],
         [ 'streak',    qw(streak_week streak_month streak_master eternal_flame) ],
         [ 'comeback',  qw(comeback_week comeback_month comeback_legend phoenix_rising) ],
+        [ 'quote contribution',   qw(archivist master_archivist) ],
+        [ 'factoid contribution', qw(lorekeeper encyclopedist) ],
     );
     my $monotone = 0;
     for my $ladder (@ladders) {